remove PySide versions

+++ src/dial.py
@@ -1,8 +1,8 @@
 from math import ceil, cos, degrees, floor, pi, sin
 
-from PySide6.QtCore import Qt
-from PySide6.QtGui import QColor, QFont, QPainter, QPalette
-from PySide6.QtWidgets import QFrame, QLabel, QWidget
+from PyQt5.QtCore import Qt
+from PyQt5.QtGui import QColor, QFont, QPainter, QPalette
+from PyQt5.QtWidgets import QFrame, QLabel, QWidget
 
 #todo: make resize event redraw everything
 #todo: make updating center label with unit
@@ -119,16 +119,16 @@ class Dial(QWidget):
         unit_dial = QFrame(frame)
         unit_dial.resize(section_x_radius + x_rad_offset - section_radius * 2,
                          section_y_radius + y_rad_offset - section_radius * 2)
-        unit_dial.move(x_rad_offset - unit_dial.frameGeometry().width() / 2,
-                       y_rad_offset - unit_dial.frameGeometry().height() / 2)
+        unit_dial.move(int(x_rad_offset - unit_dial.frameGeometry().width() / 2),
+                       int(y_rad_offset - unit_dial.frameGeometry().height() / 2))
         unit_dial.show()
 
         unit_dial_top = QFrame(frame)
         unit_dial_top.resize(unit_dial.frameGeometry().size())
         unit_dial_top.show()
         unit_dial_top.move(
-            x_rad_offset - unit_dial_top.frameGeometry().width() / 2,
-            y_rad_offset - unit_dial_top.frameGeometry().height() / 2)
+            int(x_rad_offset - unit_dial_top.frameGeometry().width() / 2),
+            int(y_rad_offset - unit_dial_top.frameGeometry().height() / 2))
 
         unit_dial_bottom = QFrame(frame)
         unit_dial_bottom.resize(unit_dial.frameGeometry().size())
@@ -141,8 +141,8 @@ class Dial(QWidget):
         unit_dial_inner_border.resize(dial_inner_border_rad,
                                       dial_inner_border_rad)
         unit_dial_inner_border.move(
-            x_rad_offset - unit_dial_inner_border.frameGeometry().width() / 2,
-            y_rad_offset - unit_dial_inner_border.frameGeometry().height() / 2)
+            int(x_rad_offset - unit_dial_inner_border.frameGeometry().width() / 2),
+            int(y_rad_offset - unit_dial_inner_border.frameGeometry().height() / 2))
         unit_dial_inner_border.show()
 
         unit_dial_mask = QFrame(frame)
